COL-1785, 'canvas' and 'socket' objects managed in Vue.prototype, not vuex store
