+++ package.json
@@ -15,7 +15,7 @@
     "@vue-a11y/announcer": "2.1.0",
     "axios": "0.26.0",
     "core-js": "3.9.1",
-    "fabric": "^5.2.1",
+    "fabric": "5.2.1",
     "highcharts": "9.1.0",
     "highcharts-vue": "1.3.5",
     "lodash": "4.17.21",
