+++ src/store/whiteboarding/listeners.ts
@@ -1,98 +1,79 @@
 import _ from 'lodash'
-import fabric from 'fabric'
 import fabricator from '@/store/whiteboarding/fabricator'
+import store from '@/store'
+import Vue from 'vue'
+import {fabric} from 'fabric'
+
+const p = Vue.prototype
 
 const addFabricCanvasListenters = (state: any) => {
   // Indicate that the currently selected elements are in the process of being moved, scaled or rotated
   const setModifyingElement = () => state.isModifyingElement = true
-  state.canvas.on('object:moving', setModifyingElement)
-  state.canvas.on('object:scaling', setModifyingElement)
-  state.canvas.on('object:rotating', setModifyingElement)
-  state.canvas.on('object:moving', (event: any) => fabricator.ensureWithinCanvas(event, state))
+  p.$canvas.on('object:moving', setModifyingElement)
+  p.$canvas.on('object:scaling', setModifyingElement)
+  p.$canvas.on('object:rotating', setModifyingElement)
+  p.$canvas.on('object:moving', (event: any) => fabricator.ensureWithinCanvas(event))
 
-  // One or multiple whiteboard state.canvas elements have been updated by the current user
-  state.canvas.on('object:modified', (event: any) => {
+  // One or multiple whiteboard canvas elements have been updated by the current user
+  p.$canvas.on('object:modified', (event: any) => {
     // Ensure that none of the modified objects are positioned off screen
-    fabricator.ensureWithinCanvas(event, state)
+    fabricator.ensureWithinCanvas(event)
     // Get the selected whiteboard elements
-    const elements = fabricator.getActiveElements(state.canvas)
+    const elements = fabricator.getActiveElements()
     // Notify the server about the updates
     fabricator.saveElementUpdates(elements, state)
   })
 
   // Indicate that the currently selected elements are no longer being modified once moving, scaling or rotating has finished
-  state.canvas.on('object:modified', () => state.isModifyingElement = false)
+  p.$canvas.on('object:modified', () => store.dispatch('whiteboarding/setIsModifyingElement', false))
 
   /**
    * Draw a box around the currently selected element(s) and use this box
    * to position the buttons that allow the selected element(s) to be modified
    */
-  state.canvas.on('after:render', () => {
-    if (!state.isModifyingElement && state.canvas.getActiveObject() || state.canvas.getActiveGroup()) {
-      // Get the bounding rectangle around the currently selected element(s)
-      let bound: any
-      if (state.canvas.getActiveObject()) {
-        bound = state.canvas.getActiveObject().getBoundingRect()
-      } else if (state.canvas.getActiveGroup()) {
-        bound = state.canvas.getActiveGroup().getBoundingRect()
-      }
-      if (bound) {
-        // Explicitly draw the bounding rectangle
-        state.canvas.contextContainer.strokeStyle = '#0295DE'
-        state.canvas.contextContainer.strokeRect(bound.left - 10, bound.top - 10, bound.width + 20, bound.height + 20)
-      }
-
-      // Position the buttons to modify the selected element(s)
-      const editButtons = document.getElementById('whiteboards-board-editelement')
-      if (editButtons) {
-        editButtons.style.left = (bound.left - 10) + 'px'
-        editButtons.style.top = (bound.top + bound.height + 15) + 'px'
-      }
-    }
-  })
+  p.$canvas.on('after:render', () => store.dispatch('whiteboarding/afterCanvasRender'))
 
   /**
    * When a new group has been added programmatically added, it needs to be programmatically
-   * removed from the state.canvas when the group is deselected
+   * removed from the canvas when the group is deselected
    */
-  state.canvas.on('before:selection:cleared', () => {
-    if (state.canvas.getActiveGroup()) {
-      state.canvas.remove(state.canvas.getActiveGroup())
+  p.$canvas.on('before:selection:cleared', () => {
+    if (p.$canvas.getActiveGroup()) {
+      p.$canvas.remove(p.$canvas.getActiveGroup())
     }
   })
 
-  // Recalculate the size of the whiteboard state.canvas when a selection has been deselected
-  state.canvas.on('selection:cleared', () => fabricator.setCanvasDimensions(state))
+  // Recalculate the size of the whiteboard canvas when a selection has been deselected
+  p.$canvas.on('selection:cleared', () => fabricator.setCanvasDimensions(state))
 
   /**
-   * A new element was added to the whiteboard state.canvas by the current user
+   * A new element was added to the whiteboard canvas by the current user
    */
-  state.canvas.on('object:added', (event: any) => {
+  p.$canvas.on('object:added', (event: any) => {
     const element = event.target
     // Don't add a new text element until text has been entered
     if (element.type === 'i-text' && !element.text.trim()) {
       return false
     }
-    // If the element already has a unique id, it was added by a different user and
-    // there is no need to persist the addition
+    // If the element already has a unique id, it was added by a different user and there is no need to persist the addition
     if (!element.get('uuid') && !element.get('isHelper')) {
-      fabricator.saveNewElement(element, state)
-      // Recalculate the size of the whiteboard state.canvas
+      fabricator.saveNewElement(element)
+      // Recalculate the size of the whiteboard canvas
       fabricator.setCanvasDimensions(state)
     }
   })
 
   /**
-   * The mouse is pressed down on the whiteboard state.canvas
+   * The mouse is pressed down on the whiteboard canvas
    */
-   state.canvas.on('mouse:down', (event: any) => {
-    // Only start drawing a shape when the state.canvas is in shape mode
+   p.$canvas.on('mouse:down', (event: any) => {
+    // Only start drawing a shape when the canvas is in shape mode
     if (state.mode === 'shape') {
       // Indicate that drawing a shape has started
       state.isDrawingShape = true
 
       // Keep track of the point where drawing the shape started
-      state.startShapePointer = state.canvas.getPointer(event.e)
+      state.startShapePointer = p.$canvas.getPointer(event.e)
 
       // Create the basic shape of the selected type that will
       // be used as the drawing guide. The originX and originY
@@ -117,18 +98,18 @@ const addFabricCanvasListenters = (state: any) => {
       // Indicate that this element is a helper element that should
       // not be saved back to the server
       state.shape.set('isHelper', true)
-      state.canvas.add(state.shape)
+      p.$canvas.add(state.shape)
     }
   })
 
   /**
-   * The mouse is moved on the whiteboard state.canvas
+   * The mouse is moved on the whiteboard canvas
    */
-  state.canvas.on('mouse:move', (event: any) => {
-    // Only continue drawing the shape when the whiteboard state.canvas is in shape mode
+  p.$canvas.on('mouse:move', (event: any) => {
+    // Only continue drawing the shape when the whiteboard canvas is in shape mode
     if (state.isDrawingShape) {
       // Get the current position of the mouse
-      const currentShapePointer = state.canvas.getPointer(event.e)
+      const currentShapePointer = p.$canvas.getPointer(event.e)
 
       // When the user has moved the cursor to the left of the original
       // starting point, move the left of the circle to that point so
@@ -159,20 +140,20 @@ const addFabricCanvasListenters = (state: any) => {
           'height': Math.abs(state.startShapePointer.y - currentShapePointer.y)
         })
       }
-      state.canvas.renderAll()
+      p.$canvas.requestRenderAll()
     }
   })
 
   /**
-   * The mouse is released on the whiteboard state.canvas
+   * The mouse is released on the whiteboard canvas
    */
-  state.canvas.on('mouse:up', () => {
+  p.$canvas.on('mouse:up', () => {
     if (state.isDrawingShape) {
       // Indicate that shape drawing has stopped
       state.isDrawingShape = false
       // Switch the toolbar back to move mode
       state.mode = 'move'
-      // Clone the drawn shape and add the clone to the state.canvas.
+      // Clone the drawn shape and add the clone to the canvas.
       // This is caused by a bug in Fabric where it initially uses
       // the size when drawing started to position the controls. Cloning
       // ensures that the controls are added in the correct position.
@@ -186,19 +167,19 @@ const addFabricCanvasListenters = (state: any) => {
       // therefore be saved back to the server
       finalShape.set('isHelper', false)
 
-      state.canvas.add(finalShape)
-      state.canvas.remove(state.shape)
+      p.$canvas.add(finalShape)
+      p.$canvas.remove(state.shape)
       // Select the added shape
-      state.canvas.setActiveObject(finalShape)
+      p.$canvas.setActiveObject(finalShape)
     }
   })
   /**
-   * Add an editable text field to the whiteboard state.canvas
+   * Add an editable text field to the whiteboard canvas
    */
-   state.canvas.on('mouse:down', function(ev) {
+   p.$canvas.on('mouse:down', function(ev) {
     if (state.mode === 'text') {
       // Add the text field to where the user clicked
-      const textPointer = state.canvas.getPointer(ev.e)
+      const textPointer = p.$canvas.getPointer(ev.e)
 
       // Start off with an empty text field
       const text = new fabric.IText('', {
@@ -208,11 +189,11 @@ const addFabricCanvasListenters = (state: any) => {
         'fontSize': state.text.selected.size,
         'fill': state.text.selected.color.color
       })
-      state.canvas.add(text)
+      p.$canvas.add(text)
 
       // Put the editable text field in edit mode straight away
       setTimeout(function() {
-        state.canvas.setActiveObject(text)
+        p.$canvas.setActiveObject(text)
         text.enterEditing()
         // The textarea needs to be put in edit mode manually
         // @see https://github.com/kangax/fabric.js/issues/1740
@@ -223,31 +204,31 @@ const addFabricCanvasListenters = (state: any) => {
 }
 
 const addSocketListeners = (state: any) => {
-  if (state.socket) {
+  if (p.$socket) {
     /**
-     * One or multiple whiteboard state.canvas elements were updated by a different user
+     * One or multiple whiteboard canvas elements were updated by a different user
      */
-    state.socket.on('updateActivity', (elements: any) => {
+    p.$socket.on('updateActivity', (elements: any) => {
       // Deactivate the current group if any of the updated elements are in the current group
-      $_deactiveActiveGroupIfOverlap(elements, state)
+      $_deactiveActiveGroupIfOverlap(elements)
       // Update the elements
       _.each(elements, function(element) {
         $_updateCanvasElement(state, element.uuid, element)
       })
-      // Recalculate the size of the whiteboard state.canvas
+      // Recalculate the size of the whiteboard canvas
       fabricator.setCanvasDimensions(state)
     })
     /**
-     * A whiteboard state.canvas element was added by a different user
+     * A whiteboard canvas element was added by a different user
      */
-     state.socket.on('addActivity', function(elements) {
+     p.$socket.on('addActivity', function(elements) {
       _.each(elements, (element: any) => {
         const callback = (e: any) => {
-          // Add the element to the whiteboard state.canvas and move it to its appropriate index
-          state.canvas.add(e)
+          // Add the element to the whiteboard canvas and move it to its appropriate index
+          p.$canvas.add(e)
           element.moveTo(e.get('index'))
-          state.canvas.renderAll()
-          // Recalculate the size of the whiteboard state.canvas
+          p.$canvas.requestRenderAll()
+          // Recalculate the size of the whiteboard canvas
           fabricator.setCanvasDimensions(state)
         }
         fabricator.deserializeElement(state, element, callback)
@@ -255,35 +236,35 @@ const addSocketListeners = (state: any) => {
     })
 
     /**
-     * One or multiple whiteboard state.canvas elements were deleted by a different user
+     * One or multiple whiteboard canvas elements were deleted by a different user
      */
-    state.socket.on('deleteActivity', function(elements) {
+    p.$socket.on('deleteActivity', function(elements) {
       // Deactivate the current group if any of the deleted elements are in the current group
-      $_deactiveActiveGroupIfOverlap(elements, state)
+      $_deactiveActiveGroupIfOverlap(elements)
       // Delete the elements
       _.each(elements, function(element) {
-        element = fabricator.getCanvasElement(state.canvas, element.uuid)
+        element = fabricator.getCanvasElement(element.uuid)
         if (element) {
-          state.canvas.remove(element)
+          p.$canvas.remove(element)
         }
       })
-      // Recalculate the size of the whiteboard state.canvas
+      // Recalculate the size of the whiteboard canvas
       fabricator.setCanvasDimensions(state)
     })
   }
 }
 
-const addModalListeners = (state: any) => {
-  // Set the toolbar back to move mode when the asset and export tooltips are hidden.
-  state.$on('tooltip.hide', function(ev, $tooltip) {
-    if ((state.mode === 'asset' && $tooltip.$id === 'whiteboards-board-asset-trigger') || (state.mode === 'export' && $tooltip.$id === 'whiteboards-board-export-trigger')) {
-      state.mode = 'move'
-    }
-  })
-  // Change the drawing color when a new color has been selected in the color picker
-  state.$watch('draw.selected.color', () => state.canvas.freeDrawingBrush.color = state.draw.selected.color.color, true)
-  // Change the drawing line width when a new line width has been selected in the width picker
-  state.$watch('draw.selected.lineWidth', () => state.canvas.freeDrawingBrush.width = parseInt(state.draw.selected.lineWidth, 10), true)
+const addModalListeners = () => {
+  // TODO: Set the toolbar back to move mode when the asset and export tooltips are hidden.
+  // state.$on('tooltip.hide', function(ev, $tooltip) {
+  //   if ((state.mode === 'asset' && $tooltip.$id === 'whiteboards-board-asset-trigger') || (state.mode === 'export' && $tooltip.$id === 'whiteboards-board-export-trigger')) {
+  //     state.mode = 'move'
+  //   }
+  // })
+  // // Change the drawing color when a new color has been selected in the color picker
+  // state.$watch('draw.selected.color', () => p.$canvas.freeDrawingBrush.color = state.draw.selected.color.color, true)
+  // // Change the drawing line width when a new line width has been selected in the width picker
+  // state.$watch('draw.selected.lineWidth', () => p.$canvas.freeDrawingBrush.width = parseInt(state.draw.selected.lineWidth, 10), true)
 }
 
 export default {
@@ -297,25 +278,25 @@ export default {
  * Deactivate the active group if any of the provided elements are a part of the active group
  * elements: The elements that should be checked for presence in the active group
  */
- const $_deactiveActiveGroupIfOverlap = (elements: any[], state: any) => {
-  const group = state.canvas.getActiveGroup()
+ const $_deactiveActiveGroupIfOverlap = (elements: any[]) => {
+  const group = p.$canvas.getActiveGroup()
   if (group) {
     const intersection = _.intersection(_.map(group.objects, 'uuid'), _.map(elements, 'uuid'))
     if (intersection.length > 0) {
-      state.canvas.discardActiveGroup().renderAll()
+      p.$canvas.discardActiveGroup().requestRenderAll()
     }
   }
 }
 
 /**
- * Update the appearance of a Fabric.js state.canvas element
+ * Update the appearance of a Fabric.js canvas element
  *
  * @param  {Number}         uuid               The id of the element to update
- * @param  {Object}         update            The updated values to apply to the state.canvas element
+ * @param  {Object}         update            The updated values to apply to the canvas element
  * @return {void}
  */
  const $_updateCanvasElement = (state: any, uuid: number, update: any): any => {
-  const element = fabricator.getCanvasElement(state.canvas, uuid)
+  const element = fabricator.getCanvasElement(uuid)
 
   const updateElementProperties = () => {
     // Update all element properties, except for the image source. The image
@@ -329,7 +310,7 @@ export default {
     // re-render the element after it has been loaded
     if (element.type === 'image' && element.getSrc() !== update.src) {
       element.setSrc(update.src, function() {
-        state.canvas.renderAll()
+        p.$canvas.requestRenderAll()
         // Ensure that the correct position is applied
         fabricator.restoreLayers(state)
       })
